able to run server locally with reload
