+++ docker-compose.yml
@@ -1,10 +1,12 @@
 version: "3.0"
 services:
-  open-operator-api:
+  open-operator-server:
     build:
       context: .
       dockerfile: ./Dockerfile.server
-    command: ["sh", "-c", "sleep 12; python ./examples/server.py"] # wait for neo4j to start
+    volumes:
+      - .:/usr/src/app
+    command: ["sh", "-c", "sleep 15; python ./server.py"] # wait for neo4j to start
     ports:
       - 8080:8080
     env_file:
