+++ openoperator/core/__init__.py
@@ -2,7 +2,6 @@ from .portfolio import Portfolio
 from .facility.facility import Facility
 from .device import DeviceManager
 from .point import PointManager
-from .server import server
 from .user import User
 from .open_operator import OpenOperator
 from .documents import Documents
